+++ src/docs/HISTORY.md
@@ -40,6 +40,7 @@ v1.4.0
   RMB->Update->Thumbnails.
 - Added thorough list of Wayland compositors with their version commands.
 - Added Desktop Envionment report when using X11 or Wayland.
+- Fixed finding all preferences with STUDIOPATH.
 
 
 v1.3.9